agent/my_agent/requirements.txt
langchain==0.3.25
langgraph==0.4.5

langgraph-cli[inmem]==0.2.10

langchain-openai==0.3.17

instructor==1.8.2
openai==1.79.0
aiohttp==3.11.18

python-dotenv==1.1.0
pydantic==2.11.4
elevenlabs==2.1.0
aiofiles==24.1.0
